allow floating point encoderCountsPerRotation
this is needed if the gear ratio is not integer

+++ Assets/Scripts/Robot/Physics.cs
@@ -8,7 +8,7 @@ public class Physics : MonoBehaviour
 {
 	public float wheelDiameterMm=43.2f;
 	public float wheelbaseMm=250.0f;
-	public int encoderCountsPerRotation=360;
+	public float encoderCountsPerRotation=360.0f;
 	public int maxEncoderCountsPerSecond=1000;
 	public bool reverseMotorPolarity=false;
